requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.2
django-multiselectfield==0.1.11
django-oauth2-provider==0.2.6.1
djangorestframework==3.11.0
djangorestframework-oauth==1.1.0
httplib2==0.15.0
idna==2.8
oauth2==1.9.0.post1
oauthlib==3.1.0
Pillow==6.2.1
psycopg2==2.8.4
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==0.3.1
shortuuid==0.5.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
